Update DimensionTask submit
Checks the dimension before counting it to the player. Initially, the player could simply jump over the progress, ignoring the quest line (for example, if the quest has a dependency).

However, even if the quest is credited to the player ignoring the quest line -> everything starts to work incorrectly. The player cannot complete further quests due to the fact that the dimension was counted to him before he could open it. And all this results in the fact that the player simply cannot continue to study quests - they simply do not work and get stuck on the dimension.

I fixed this issue for my server and decided to report it to you.